Extract methods to interfaces and provide dependencies
