+++ app/src/main/java/space/taran/arkmemo/di/PreferencesModule.kt
@@ -1,0 +1,4 @@
+package space.taran.arkmemo.di
+
+class PreferencesModule {
+}
