+++ app/src/main/java/space/taran/arkmemo/di/RepositoryModule.kt
@@ -1,0 +1,4 @@
+package space.taran.arkmemo.di
+
+class RepositoryModule {
+}
